fixed mod flag, added chmod

+++ kernel/file.h
@@ -27,7 +27,7 @@ struct inode {
   short nlink;
   uint size;
   uint addrs[NDIRECT+1];
-  uint8 mod; //mod
+  short mod; //mod
 };
 
 // map major device number to device functions.
